// app/validation/schemas/calender.ts
import { z } from "zod";

export const calendarSchema = z.object({
  title: z.string().min(3, "Title must be at least 3 characters"),
  description: z.string().min(10, "Description must be at least 10 characters"),
  date: z.string().min(1, "Date is required"),
  time: z.string().min(1, "Time is required"),
  location: z.string().min(1, "Location is required"),
});
export type CalendarFormData = z.infer<typeof calendarSchema>;